Conflicts:
	besch/writer/bridge_writer.cc
	besch/writer/root_writer.cc
	besch/writer/tunnel_writer.cc
	besch/writer/vehicle_writer.cc
	besch/writer/way_obj_writer.cc
	besch/writer/way_writer.cc
	dataobj/loadsave.h
	dataobj/translator.cc
	dataobj/translator.h
	gui/map_frame.cc
	gui/pakselector.cc
	gui/savegame_frame.cc
	player/simplay.cc
	simcity.h

+++ simconvoi.cc
@@ -2963,7 +2963,7 @@ void convoi_t::get_freight_info(cbuffer_t & buf)
 		ALLOCA(uint32, max_loaded_waren, n);
 		MEMZERON(max_loaded_waren, n);
 
-		for (size_t i = 0; i != anz_vehikel; ++i) {
+		for(  uint32 i = 0;  i != anz_vehikel;  ++i  ) {
 			const vehikel_t* v = fahr[i];
 
 			// first add to capacity indicator
